apps/api/pyproject.toml
[project]
name = "llmbench-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "asyncpg>=0.29",
    "redis>=5.0",
    "pydantic-settings>=2.2",
    "argon2-cffi>=23.1",
    "structlog>=24.1",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "email-validator>=2.1",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "anyio[trio]>=4.3", "httpx>=0.27", "pytest-asyncio>=0.23"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
